Revert and move into site editor where we're unregistering the core command
